+++ examples/example 1/app.py
@@ -1,22 +1,42 @@
-from flask import Flask, jsonify, g
-from .filters import UserInputFilter, ProfileInputFilter, ProductInputFilter
+from flask import Flask, Response, g
+
+from .filters import ProductInputFilter, ProfileInputFilter, UserInputFilter
 
 app = Flask(__name__)
 
-@app.route('/api/user', methods=['POST'], endpoint='create-user')
+
+@app.route("/api/user", methods=["POST"], endpoint="create-user")
 @UserInputFilter.validate()
 def create_user():
-    return jsonify({"message": "User created successfully", "data": g.validated_data}), 201
+    return Response(
+        {"message": "User created successfully", "data": g.validated_data},
+        201,
+    )
 
-@app.route('/api/profile', methods=['POST'], endpoint='create-profile')
+
+@app.route("/api/profile", methods=["POST"], endpoint="create-profile")
 @ProfileInputFilter.validate()
 def create_profile():
-    return jsonify({"message": "Profile created successfully", "data": g.validated_data}), 201
+    return Response(
+        {
+            "message": "Profile created successfully",
+            "data": g.validated_data,
+        },
+        201,
+    )
 
-@app.route('/api/product', methods=['POST'], endpoint='create-product')
+
+@app.route("/api/product", methods=["POST"], endpoint="create-product")
 @ProductInputFilter.validate()
 def create_products():
-    return jsonify({"message": "Products created successfully", "data": g.validated_data}), 201
+    return Response(
+        {
+            "message": "Products created successfully",
+            "data": g.validated_data,
+        },
+        201,
+    )
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     app.run(debug=True)