+++ examples/example 1/filters/user_inputfilter.py
@@ -6,26 +6,8 @@ class UserInputFilter(InputFilter):
     def __init__(self):
         super().__init__()
 
-        self.add(
-            'name',
-            required=True,
-            validators=[
-                IsStringValidator()
-            ]
-        )
+        self.add("name", required=True, validators=[IsStringValidator()])
 
-        self.add(
-            'age',
-            required=True,
-            validators=[
-                IsIntegerValidator()
-            ]
-        )
+        self.add("age", required=True, validators=[IsIntegerValidator()])
 
-        self.add(
-            'email',
-            required=True,
-            validators=[
-                IsStringValidator()
-            ]
-        )
+        self.add("email", required=True, validators=[IsStringValidator()])